:art: refactor: 코드 메소드화
